Service entidade InventoryTransaction implementado

+++ src/main/java/com/restaurant/restaurantManagement/model/dto/inventoryTransaction/CreateInventoryTransactionDTO.java
@@ -17,7 +17,7 @@ public record CreateInventoryTransactionDTO(
 
         @NotNull
         @Positive
-        int quantity,
+        Long quantity,
 
         @NotNull
         @Positive
